+++ app/src/main/java/de/triology/cas/oidc/config/CesOidcConfiguration.java
@@ -8,6 +8,7 @@ import de.triology.cas.oidc.beans.delegation.CesCustomDelegatedAuthenticationCli
 import de.triology.cas.oidc.beans.delegation.CesDelegatedAuthenticationPreProcessor;
 import de.triology.cas.oidc.beans.delegation.CesDelegatedClientUserProfileProvisioner;
 import lombok.extern.slf4j.Slf4j;
+import org.apache.commons.lang.StringUtils;
 import org.apereo.cas.authentication.principal.DelegatedAuthenticationPreProcessor;
 import org.apereo.cas.config.OidcConfiguration;
 import org.apereo.cas.configuration.CasConfigurationProperties;
@@ -27,6 +28,7 @@ import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.webflow.execution.Action;
 
+import java.util.Arrays;
 import java.util.List;
 
 @Configuration("CesOidcConfiguration")
@@ -43,6 +45,15 @@ public class CesOidcConfiguration {
 
     @Value("${ces.delegation.oidc.attributeMapping:#{\"\"}}")
     private String attributesMappingsString;
+
+    @Value("${ces.delegation.oidc.allowedGroups:#{\"\"}}")
+    private String allowedGroupsConfigString;
+
+    @Value("${ces.delegation.oidc.initialAdminUsernames:#{\"\"}}")
+    private String initialAadminUsernamesConfigString;
+
+    @Value("${ces.delegation.oidc.adminGroups:#{\"\"}}")
+    private String adminGroupsConfigString;
 
     @Bean
     @RefreshScope
@@ -65,8 +76,10 @@ public class CesOidcConfiguration {
     @RefreshScope
     public DelegatedClientUserProfileProvisioner clientUserProfileProvisioner(final CasConfigurationProperties casProperties) {
         UserManager userManager = getUserManager(casProperties);
+        String[] initialAdminUsernames = splitAndTrim(initialAadminUsernamesConfigString);
+        String[] adminGroups = splitAndTrim(adminGroupsConfigString);
 
-        return new CesDelegatedClientUserProfileProvisioner(userManager);
+        return new CesDelegatedClientUserProfileProvisioner(userManager, initialAdminUsernames, adminGroups);
     }
 
     @Bean
@@ -74,8 +87,9 @@ public class CesOidcConfiguration {
     public DelegatedAuthenticationPreProcessor delegatedAuthenticationPreProcessor(final CasConfigurationProperties casProperties) {
         UserManager userManager = getUserManager(casProperties);
         List<AttributeMapping> attributeMappings = AttributeMapping.fromPropertyString(attributesMappingsString);
+        String[] allowedGroups = splitAndTrim(allowedGroupsConfigString);
 
-        return new CesDelegatedAuthenticationPreProcessor(attributeMappings, userManager);
+        return new CesDelegatedAuthenticationPreProcessor(userManager, attributeMappings, allowedGroups);
     }
 
     private static UserManager getUserManager(CasConfigurationProperties casProperties) {
@@ -83,5 +97,15 @@ public class CesOidcConfiguration {
         PooledConnectionFactory connectionFactory = LdapUtils.newLdaptivePooledConnectionFactory(ldapProperties);
 
         return new UserManager(ldapProperties.getBaseDn(), new LdapOperationFactory(connectionFactory));
+    }
+
+    private static String[] splitAndTrim(String input) {
+        if (StringUtils.isBlank(input)) {
+            return new String[0];
+        }
+
+        String[] result = StringUtils.split(input, ",");
+
+        return Arrays.stream(result).map(String::trim).toArray(String[]::new);
     }
 }