+++ app/src/test/java/de/triology/cas/oidc/beans/delegation/CesDelegatedClientUserProfileProvisionerTest.java
@@ -3,6 +3,7 @@ package de.triology.cas.oidc.beans.delegation;
 import de.triology.cas.ldap.CesInternalLdapUser;
 import de.triology.cas.ldap.CesLdapException;
 import de.triology.cas.ldap.UserManager;
+import org.apereo.cas.authentication.principal.Principal;
 import org.junit.Test;
 import org.mockito.Mockito;
 import org.pac4j.core.profile.CommonProfile;
@@ -10,6 +11,9 @@ import org.pac4j.core.profile.UserProfile;
 import org.pac4j.oidc.profile.OidcProfile;
 import org.pac4j.oidc.profile.OidcProfileDefinition;
 
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
 import java.util.Set;
 
 import static org.junit.jupiter.api.Assertions.*;
@@ -33,7 +37,7 @@ public class CesDelegatedClientUserProfileProvisionerTest {
         CesInternalLdapUser expectedCesUser = new CesInternalLdapUser("test_user", "Test", "User", "Test User", "test@user.de", true);
         Mockito.doNothing().when(userManagerMock).updateUser(expectedCesUser);
 
-        CesDelegatedClientUserProfileProvisioner provisioner = new CesDelegatedClientUserProfileProvisioner(userManagerMock);
+        CesDelegatedClientUserProfileProvisioner provisioner = new CesDelegatedClientUserProfileProvisioner(userManagerMock, null, null);
         provisioner.execute(null, profile, null, null);
 
         verify(userManagerMock, times(1)).updateUser(expectedCesUser);
@@ -54,10 +58,69 @@ public class CesDelegatedClientUserProfileProvisionerTest {
         CesInternalLdapUser expectedCesUser = new CesInternalLdapUser("test_user", "Test", "User", "Test User", "test@user.de", true);
         Mockito.doNothing().when(userManagerMock).createUser(expectedCesUser);
 
-        CesDelegatedClientUserProfileProvisioner provisioner = new CesDelegatedClientUserProfileProvisioner(userManagerMock);
-        provisioner.execute(null, profile, null, null);
+        Principal principal = mock(Principal.class);
+        when(principal.getAttributes()).thenReturn(new HashMap<>());
+
+        CesDelegatedClientUserProfileProvisioner provisioner = new CesDelegatedClientUserProfileProvisioner(userManagerMock, null, null);
+        provisioner.execute(principal, profile, null, null);
 
         verify(userManagerMock, times(1)).createUser(expectedCesUser);
+    }
+
+    @Test
+    public void testExecute_withNewUserAddAdminGroups() throws Throwable {
+        OidcProfile profile = new OidcProfile();
+        profile.addAttribute(OidcProfileDefinition.PREFERRED_USERNAME, "test_user");
+        profile.addAttribute(OidcProfileDefinition.GIVEN_NAME, "Test");
+        profile.addAttribute(OidcProfileDefinition.FAMILY_NAME, "User");
+        profile.addAttribute(OidcProfileDefinition.NAME, "Test User");
+        profile.addAttribute(OidcProfileDefinition.EMAIL, "test@user.de");
+
+        UserManager userManagerMock = mock(UserManager.class);
+        when(userManagerMock.getUserByUid("test_user")).thenReturn(null);
+
+        CesInternalLdapUser expectedCesUser = new CesInternalLdapUser("test_user", "Test", "User", "Test User", "test@user.de", true);
+        Mockito.doNothing().when(userManagerMock).createUser(expectedCesUser);
+        Mockito.doNothing().when(userManagerMock).addUserToGroup(expectedCesUser, "group1");
+        Mockito.doNothing().when(userManagerMock).addUserToGroup(expectedCesUser, "group2");
+
+        Principal principal = mock(Principal.class);
+        when(principal.getAttributes()).thenReturn(new HashMap<>());
+
+        CesDelegatedClientUserProfileProvisioner provisioner = new CesDelegatedClientUserProfileProvisioner(userManagerMock, new String[]{"test_user"}, new String[]{"group1", "group2"});
+        provisioner.execute(principal, profile, null, null);
+
+        verify(userManagerMock, times(1)).createUser(expectedCesUser);
+        verify(userManagerMock, times(1)).addUserToGroup(expectedCesUser, "group1");
+        verify(userManagerMock, times(1)).addUserToGroup(expectedCesUser, "group2");
+    }
+
+    @Test
+    public void testExecute_withNewUserDoNotAddAdminGroups() throws Throwable {
+        OidcProfile profile = new OidcProfile();
+        profile.addAttribute(OidcProfileDefinition.PREFERRED_USERNAME, "test_user");
+        profile.addAttribute(OidcProfileDefinition.GIVEN_NAME, "Test");
+        profile.addAttribute(OidcProfileDefinition.FAMILY_NAME, "User");
+        profile.addAttribute(OidcProfileDefinition.NAME, "Test User");
+        profile.addAttribute(OidcProfileDefinition.EMAIL, "test@user.de");
+
+        UserManager userManagerMock = mock(UserManager.class);
+        when(userManagerMock.getUserByUid("test_user")).thenReturn(null);
+
+        CesInternalLdapUser expectedCesUser = new CesInternalLdapUser("test_user", "Test", "User", "Test User", "test@user.de", true);
+        Mockito.doNothing().when(userManagerMock).createUser(expectedCesUser);
+        Mockito.doNothing().when(userManagerMock).addUserToGroup(expectedCesUser, "group1");
+        Mockito.doNothing().when(userManagerMock).addUserToGroup(expectedCesUser, "group2");
+
+        Principal principal = mock(Principal.class);
+        when(principal.getAttributes()).thenReturn(new HashMap<>());
+
+        CesDelegatedClientUserProfileProvisioner provisioner = new CesDelegatedClientUserProfileProvisioner(userManagerMock, new String[]{"super_admin"}, new String[]{"group1", "group2"});
+        provisioner.execute(principal, profile, null, null);
+
+        verify(userManagerMock, times(1)).createUser(expectedCesUser);
+        verify(userManagerMock, times(0)).addUserToGroup(expectedCesUser, "group1");
+        verify(userManagerMock, times(0)).addUserToGroup(expectedCesUser, "group2");
     }
 
     @Test
@@ -66,7 +129,7 @@ public class CesDelegatedClientUserProfileProvisionerTest {
 
         UserManager userManagerMock = mock(UserManager.class);
 
-        CesDelegatedClientUserProfileProvisioner provisioner = new CesDelegatedClientUserProfileProvisioner(userManagerMock);
+        CesDelegatedClientUserProfileProvisioner provisioner = new CesDelegatedClientUserProfileProvisioner(userManagerMock, null, null);
 
         RuntimeException e = assertThrows(RuntimeException.class, () -> provisioner.execute(null, profile, null, null));
         assertEquals("Unsupported profile type: CommonProfile", e.getMessage());