+++ CHANGELOG.md
@@ -7,6 +7,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [v7.0.8-10] - 2025-01-10
+### Added
+- Add configuration for `allowed_groups` and `initial_admin_user` in delegated authentication
+
+### Changed
+- [#246] Update base image to Alpine 3.21.0 and Java to 21.0.5-p11
+- Update Tomcat to 10.1.34
+
+### Fixed
+- [#246] Fix a restart loop if the config key `oidc/enabled` was not set.
+
 ## [v7.0.8-9] - 2024-12-20
 ### Added
 - Add http health-check, so that the dogu will get healthy when the start of the web-application is completed